third_party/WebKit/LayoutTests/inspector/console/console-uncaught-promise-expected.txt

CONSOLE ERROR: line 23: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 44: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 62: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
CONSOLE ERROR: line 68: Uncaught (in promise) 42
CONSOLE ERROR: line 73: Uncaught (in promise) 1e+100
CONSOLE ERROR: line 78: Uncaught (in promise) foo
CONSOLE ERROR: line 83: Uncaught (in promise) #<Object>
CONSOLE ERROR: line 88: Uncaught (in promise) undefined
CONSOLE ERROR: line 93: Uncaught (in promise) SecurityError: Failed to register a ServiceWorker: The URL protocol of the current origin ('') is not supported.
Tests that uncaught promise rejections are logged into console.

console-uncaught-promise.html:23 Uncaught (in promise) Error: err1
promiseTest1 @ console-uncaught-promise.html:23
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:44 Uncaught (in promise) Error: err2
promiseTest2 @ console-uncaught-promise.html:44
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:62 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
throwDOMException @ console-uncaught-promise.html:62
catcher @ console-uncaught-promise.html:54
console-uncaught-promise.html:68 Uncaught (in promise) 42
promiseTest4 @ console-uncaught-promise.html:68
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:73 Uncaught (in promise) 1e+100
promiseTest5 @ console-uncaught-promise.html:73
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:78 Uncaught (in promise) foo
promiseTest6 @ console-uncaught-promise.html:78
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:83 Uncaught (in promise) Object {foo: 42}
    foo: 42
    __proto__: Object
promiseTest7 @ console-uncaught-promise.html:83
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:88 Uncaught (in promise) undefined
promiseTest8 @ console-uncaught-promise.html:88
timeout @ console-uncaught-promise.html:16
console-uncaught-promise.html:93 Uncaught (in promise) DOMException: Failed to register a ServiceWorker: The URL protocol of the current origin ('file://') is not supported.
promiseTest9 @ console-uncaught-promise.html:93
timeout @ console-uncaught-promise.html:16
